% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maf.R
\name{compute_maf}
\alias{compute_maf}
\title{Compute Minor Allele Frequency (MAF)}
\usage{
compute_maf(X)
}
\arguments{
\item{X}{A genotype matrix (taxa in rows, SNPs in columns)}
}
\value{
A numeric vector of MAF values for each SNP
}
\description{
This function calculates the Minor Allele Frequency (MAF) for each SNP in the genotype matrix.
It ensures that all columns are numeric before computing allele frequencies.
}
