% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Generate Manhattan Plot}
\usage{
manhattan_plot(pvals, myGM)
}
\arguments{
\item{pvals}{Vector of p-values from GWAS.}

\item{myGM}{Data frame with SNP names, chromosome, and position.}
}
\description{
This function creates a Manhattan plot for visualizing GWAS results.
}
