% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_adjustment.R
\name{pca_adjustment}
\alias{pca_adjustment}
\title{PCA Adjustment Function}
\usage{
pca_adjustment(X, C = NULL, npcs = 5)
}
\arguments{
\item{X}{Genotype matrix.}

\item{C}{Covariate matrix (optional).}

\item{npcs}{Number of principal components to include.}
}
\value{
List with adjusted covariates.
}
\description{
Performs PCA on the genotype matrix and returns adjusted covariates.
}
