% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{qq_plot}
\alias{qq_plot}
\title{Generate QQ Plot}
\usage{
qq_plot(pvals)
}
\arguments{
\item{pvals}{Vector of p-values from GWAS.}
}
\description{
This function creates a QQ plot to assess the distribution of p-values.
}
